Tree view demo created
