+++ package.json
@@ -73,6 +73,7 @@
   },
   "dependencies": {
     "@date-io/date-fns": "^1.1.0",
+    "@material-ui/icons": "^4.9.1",
     "@material-ui/pickers": "^3.2.2",
     "classnames": "^2.2.6",
     "date-fns": "^2.0.0-alpha.27",
