+++ src/components/m-table-filter-row.js
@@ -91,7 +91,8 @@ class MTableFilterRow extends React.Component {
           startAdornment: (
             <InputAdornment position="start">
               <Tooltip title={localization.filterTooltip}>
-                <this.props.icons.Filter />
+                <Icon></Icon>
+                {/* <this.props.icons.Filter /> */}
               </Tooltip>
             </InputAdornment>
           )
